Used Player from Main
