--- Assets/Scripts/VehicleController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7137c8fe61abd1e4fb5e5a827c2d5378
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
